+++ ladder-api/src/main/kotlin/kr/mashup/ladder/auth/service/AnonymousAuthService.kt
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
-class GuestAuthService(
+class AnonymousAuthService(
     private val memberRepository: MemberRepository,
 ) {
