+++ ladder-api/src/main/kotlin/kr/mashup/ladder/member/service/MemberService.kt
@@ -1,7 +1,7 @@
 package kr.mashup.ladder.member.service
 
 import kr.mashup.ladder.domain.member.domain.Member
-import kr.mashup.ladder.domain.member.domain.MemberNotFoundException
+import kr.mashup.ladder.domain.member.exception.MemberNotFoundException
 import kr.mashup.ladder.domain.member.infra.jpa.MemberRepository
 import kr.mashup.ladder.member.dto.request.UpdateMemberInfoRequest
 import kr.mashup.ladder.member.dto.response.MemberInfoResponse
