+++ ladder-api/src/main/kotlin/kr/mashup/ladder/room/service/RoomInviteService.kt
@@ -1,7 +1,7 @@
 package kr.mashup.ladder.room.service
 
 import kr.mashup.ladder.domain.room.domain.Room
-import kr.mashup.ladder.domain.room.domain.RoomNotFoundException
+import kr.mashup.ladder.domain.room.exception.RoomNotFoundException
 import kr.mashup.ladder.domain.room.infra.jpa.RoomRepository
 import kr.mashup.ladder.room.dto.response.RoomInfoResponse
 import org.springframework.stereotype.Service
