+++ ladder-api/src/test/kotlin/kr/mashup/ladder/room/RoomAcceptanceTest.kt
@@ -16,9 +16,9 @@ import kr.mashup.ladder.config.ws.WS_DESTINATION_PREFIX_QUEUE
 import kr.mashup.ladder.config.ws.WS_DESTINATION_PREFIX_TOPIC
 import kr.mashup.ladder.config.ws.WS_ENDPOINT
 import kr.mashup.ladder.config.ws.WS_USER_DESTINATION_PREFIX
-import kr.mashup.ladder.domain.common.error.ErrorCode
+import kr.mashup.ladder.domain.common.exception.ErrorCode
 import kr.mashup.ladder.domain.room.domain.emoji.EmojiType
-import kr.mashup.ladder.domain.util.JsonUtil
+import kr.mashup.ladder.domain.common.util.JsonUtil
 import kr.mashup.ladder.room.RoomFixture.Companion.`방 생성 요청값`
 import kr.mashup.ladder.room.RoomFixture.Companion.`방 재생목록 항목 신청 승인 요청값`
 import kr.mashup.ladder.room.RoomFixture.Companion.`방 재생목록 항목 신청 요청값`
