+++ ladder-domain/src/main/kotlin/kr/mashup/ladder/domain/common/util/JsonUtil.kt
@@ -1,4 +1,4 @@
-package kr.mashup.ladder.domain.util
+package kr.mashup.ladder.domain.common.util
 
 import com.fasterxml.jackson.core.type.TypeReference
 import com.fasterxml.jackson.databind.ObjectMapper
