+++ ladder-domain/src/main/kotlin/kr/mashup/ladder/domain/room/exception/RoomConflictException.kt
@@ -1,7 +1,7 @@
-package kr.mashup.ladder.domain.room.domain
+package kr.mashup.ladder.domain.room.exception
 
-import kr.mashup.ladder.domain.common.error.ErrorCode
-import kr.mashup.ladder.domain.common.error.model.LadderBaseException
+import kr.mashup.ladder.domain.common.exception.ErrorCode
+import kr.mashup.ladder.domain.common.exception.model.LadderBaseException
 
 data class RoomConflictException(
     override val message: String,
