+++ ladder-search/src/main/kotlin/kr/mashup/ladder/youtube/model/YoutubeVideoCategory.kt
@@ -1,4 +1,4 @@
-package kr.mashup.ladder.youtube.external.dto.type
+package kr.mashup.ladder.youtube.model
 
 /**
  * https://gist.github.com/dgp/1b24bf2961521bd75d6c
